Show pre-recorded gesture with Shift-[Character].
I ended up having to capture the entire Leap::Frame rather than hand to be able to replay the hand. That meant I had to serialize it in the gestures file. I currently have all the different gestures' serialized frame in a separate file but that will change in the next check in.

By the same token, I also fixed a major bug whereby since the joint positions where written to file without separation, re-reading them later on ended up on streaming in incorrect floats.